requirements.txt
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
dacite==1.9.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pyparsing==3.0.9
pyrfc6266==1.0.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soundcloud-v2==1.6.0
typing_extensions==4.12.2
urllib3==2.3.0
xattr==1.1.4